+++ lib/src/common/config.dart
@@ -4,6 +4,9 @@ const klutterPubVersion = "0.3.0";
 /// The version of the Klutter Gradle Plugin.
 const klutterGradleVersion = "2023.1.1.beta";
 
+/// The version of Kotlin to be used.
+const kotlinVersion = "1.8.20";
+
 /// The minimum SDK version for Android.
 const androidMinSdk = 21;