+++ lib/src/consumer/consumer.dart
@@ -25,4 +25,3 @@
 library consumer;
 
 export "android.dart";
-export "ios.dart";