+++ test/src/producer/android_test.dart
@@ -35,6 +35,7 @@ void main() {
         () => writeBuildGradleFile(
             pluginVersion: pluginVersion,
             packageName: packageName,
+            klutterBomVersion: "2023.1.1",
             pathToAndroid: "fake"),
         throwsA(predicate((e) =>
             e is KlutterException &&
@@ -56,6 +57,7 @@ void main() {
               pluginVersion: pluginVersion,
               packageName: packageName,
               pathToAndroid: android.path,
+          klutterBomVersion: "2023.1.1",
             ),
         throwsA(predicate((e) =>
             e is KlutterException &&
@@ -80,6 +82,7 @@ void main() {
       pluginVersion: pluginVersion,
       packageName: packageName,
       pathToAndroid: android.path,
+      klutterBomVersion: "2023.1.1",
     );
 
     expect(
@@ -97,14 +100,15 @@ void main() {
             repositories {
                 google()
                 mavenCentral()
+                mavenLocal()
                 maven { url = uri("https://repsy.io/mvn/buijs-dev/klutter") }
             }
         
             dependencies {
-                classpath platform("dev.buijs.klutter:bom:$klutterGradleVersion")
+                classpath platform("dev.buijs.klutter:bom:2023.1.1")
                 classpath "dev.buijs.klutter:gradle"
                 classpath 'com.android.tools.build:gradle:7.0.4'
-                classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.10"
+                classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.8.20"
             }
         }
         
@@ -112,6 +116,7 @@ void main() {
             repositories {
                 google()
                 mavenCentral()
+                mavenLocal()
                 maven { url = uri("https://repsy.io/mvn/buijs-dev/klutter") }
             }
         }
@@ -138,15 +143,16 @@ void main() {
         }
         
         klutter {
+            root = file("../")
             include("annotations")
             include("kore")
             include("kompose")
-            include("flutter-engine-android")
+            include("embedded")
         }
         
         dependencies {
             runtimeOnly "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.4"
-            implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.6.10"
+            implementation "org.jetbrains.kotlin:kotlin-stdlib:1.8.20"
             implementation project(":klutter:example_plugin")
         }